+++ packages/ui/theme-builder/app/utils/generate-colors.ts
